+++ R/margot_plot.R
@@ -30,8 +30,8 @@
 margot_plot <- function(.data,
                         type = c("RD", "RR"),
                         order = c("default", "alphabetical"),
-                        title,
-                        subtitle,
+                        title = NULL,
+                        subtitle = NULL,
                         estimate_scale = 1,
                         base_size = 11,
                         text_size = 2.75,
@@ -55,7 +55,6 @@ margot_plot <- function(.data,
   if (!"Estimate" %in% names(.data) || !"outcome" %in% names(.data)) {
     .data <- group_tab(.data, type = type, order = order)
   }
-
 
   # Dynamic theme adjustment
   if (is.null(plot_theme)) {
@@ -85,7 +84,12 @@ margot_plot <- function(.data,
     geom_point(size = point_size, position = position_dodge(width = 0.3)) +
     geom_vline(xintercept = if(type == "RR") 1 else 0, linetype = "solid") +
     scale_color_manual(values = c("positive" = "dodgerblue", "not reliable" = "black", "negative" = "orange")) +
-    labs(x = paste0("Causal ", ifelse(type == "RR", "risk ratio", "difference"), " scale"), y = NULL, title = title, subtitle = subtitle) +
+    labs(
+      x = paste0("Causal ", ifelse(type == "RR", "risk ratio", "difference"), " scale"),
+      y = NULL,
+      title = ifelse(is.null(title), "", title),
+      subtitle = ifelse(is.null(subtitle), "", subtitle)
+    ) +
     geom_text(aes(x = x_offset * estimate_scale, label = estimate_lab), size = text_size, hjust = 0, fontface = "bold") +
     coord_cartesian(xlim = c(x_lim_lo, x_lim_hi)) +
     plot_theme +
@@ -113,6 +117,7 @@ margot_plot <- function(.data,
 }
 # margot_plot <- function(.data,
 #                         type = c("RD", "RR"),
+#                         order = c("default", "alphabetical"),
 #                         title,
 #                         subtitle,
 #                         estimate_scale = 1,
@@ -127,56 +132,49 @@ margot_plot <- function(.data,
 #                         x_lim_lo = ifelse(type == "RR", .1, -1.75),
 #                         x_lim_hi = ifelse(type == "RR", 2.5, 1),
 #                         linewidth = .5,
-#                         plot_theme = NULL){
-#   type <- match.arg(type)
+#                         plot_theme = NULL) {
+#   require("ggplot2")
+#   require("dplyr")
 #
-#   # dynamic theme adjustment
+#   type <- match.arg(type)
+#   order <- match.arg(order)
+#
+#   # Check if the data needs processing by group_tab
+#   if (!"Estimate" %in% names(.data) || !"outcome" %in% names(.data)) {
+#     .data <- group_tab(.data, type = type, order = order)
+#   }
+#
+#
+#   # Dynamic theme adjustment
 #   if (is.null(plot_theme)) {
 #     plot_theme <- theme_classic(base_size = base_size)
 #   } else {
-#     # Apply base_size to the plot_theme
 #     plot_theme <- plot_theme + theme(text = element_text(size = base_size))
 #   }
 #
-#
-#   # Copy data to avoid direct manipulation
-#   plot_data <- .data
-#   # Define reliability based on type
-#   reliability_condition <- if (type == "RR") {
-#     list(condition = c(1, 1), label = "Causal risk ratio scale")
-#   } else {
-#     list(condition = c(0, 0), label = "Causal difference scale")
-#   }
-#
-#   plot_data$Reliability <- ifelse(
-#     plot_data$`2.5 %` > reliability_condition$condition[1] & plot_data$`97.5 %` > reliability_condition$condition[2],
-#     "positive",
-#     ifelse(
-#       plot_data$`2.5 %` < reliability_condition$condition[1] & plot_data$`97.5 %` < reliability_condition$condition[2],
-#       "negative",
-#       "zero_crossing"
-#     )
-#   )
+#   # Prepare the data for plotting, including ordering
+#   effect_size_col <- if (type == "RR") "E[Y(1)]/E[Y(0)]" else "E[Y(1)]-E[Y(0)]"
+#   .data <- .data %>%
+#     mutate(outcome = factor(outcome, levels = if (order == "alphabetical") sort(unique(outcome)) else unique(outcome))) %>%
+#     arrange(if (order == "alphabetical") outcome else desc(!!sym(effect_size_col)))
 #
 #   # Start building the plot
-#   x_name <- paste0("E[Y(1)]", ifelse(type == "RR", "/", "-"), "E[Y(0)]")
 #   out <- ggplot(
-#     data = plot_data,
+#     data = .data,
 #     aes(
-#       y = reorder(outcome, .data[[x_name]]),
-#       x = .data[[x_name]],
-#       xmin = .data$`2.5 %`,
-#       xmax = .data$`97.5 %`,
-#       group = Estimate,
-#       color = Reliability
+#       y = outcome,
+#       x = !!sym(effect_size_col),
+#       xmin = `2.5 %`,
+#       xmax = `97.5 %`,
+#       color = Estimate
 #     )
-#   ) + geom_errorbarh(aes(color = Reliability), height = .3,
+#   ) + geom_errorbarh(aes(color = Estimate), height = .3,
 #                      linewidth = linewidth, position = position_dodge(width = .3)) +
 #     geom_point(size = point_size, position = position_dodge(width = 0.3)) +
 #     geom_vline(xintercept = if(type == "RR") 1 else 0, linetype = "solid") +
-#     scale_color_manual(values = c("positive" = "dodgerblue", "zero_crossing" = "black", "negative" = "orange")) +
-#     labs(x = reliability_condition$label, y = NULL, title = title, subtitle = subtitle) +
-#     geom_text(aes(x = x_offset * estimate_scale, label = estimate_lab), size = text_size, hjust = 0, fontface = ifelse(plot_data$Estimate == "unreliable", "plain", "bold")) +
+#     scale_color_manual(values = c("positive" = "dodgerblue", "not reliable" = "black", "negative" = "orange")) +
+#     labs(x = paste0("Causal ", ifelse(type == "RR", "risk ratio", "difference"), " scale"), y = NULL, title = title, subtitle = subtitle) +
+#     geom_text(aes(x = x_offset * estimate_scale, label = estimate_lab), size = text_size, hjust = 0, fontface = "bold") +
 #     coord_cartesian(xlim = c(x_lim_lo, x_lim_hi)) +
 #     plot_theme +
 #     theme(
@@ -191,7 +189,7 @@ margot_plot <- function(.data,
 #       plot.margin = margin(t = 10, r = 10, b = 10, l = 10, unit = "pt")
 #     )
 #
-#   # Conditionally add x-axis scale modifications
+#   # Conditionally add x-axis scale modifications for RR
 #   if (type == "RR") {
 #     custom_x_labels <- function(x) {
 #       ifelse(x < 0, "", as.character(x))
@@ -201,4 +199,3 @@ margot_plot <- function(.data,
 #
 #   return(out)
 # }
-